initial sharing of stuff I had on my computer
The .tflite is the best "tensor flow lite" model I've tested.  (The other, newer, ones on the -RC phone are untested and probably worse.)  The 1130 means I made it on November 30.
snowbot_offline_blocks_editor is based on the snowbot phones and FRENZYright...zip is based on the main robot phones.  Unpack and run with Chrome browser to edit without the actual hardware available.  You can "download" blocks op-modes from the  offline blocks editor and "upload" them to the real phones later.
Power1.blk is the block code we drove on Monday afternoon.